Fix displaying edit box fields.
If an item was edited, not created as a new one, not all fields
were visible in the InnerEditBox.

Now no matter if the item is beeing edited, or beeing created
all possible fields to be set are visible.

It caused an issue, that user could not modify the existing fields.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>