Display all EditBox fields and fix saving them
There was a bug that caused missing undefined fields after loading
the XML file, and as a result, the user could not make
an extension to this cluster properly.

Also fixed an issue where some fields should be conditionally
excluded from the XML file because they cause failures in
the data model.

Reverted array types. If the array type is selected, the storage
will be fixed to `External` and user must implement it.

All attributes that uses custom Structures and Enums must
be added to the zcl.json file.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>